requirements.txt
asgiref==3.6.0
Django==4.1.7
django-debug-toolbar==4.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.3
